@@ -33,20 +33,47 @@
 	</developers>
 
 	<properties>
-		<kernel-dataaccess-hibernate.version>1.2.0.1-B1</kernel-dataaccess-hibernate.version>
-		<kernel-websubclient-api.version>1.2.0.1-B1</kernel-websubclient-api.version>
-		<kernel.version>1.2.0.1-B1</kernel.version>
-		<maven.compiler.source>11</maven.compiler.source>
- 		<maven.compiler.target>11</maven.compiler.target>
+		<kernel-dataaccess-hibernate.version>1.2.1-java21-SNAPSHOT</kernel-dataaccess-hibernate.version>
+		<kernel-websubclient-api.version>1.2.1-java21-SNAPSHOT</kernel-websubclient-api.version>
+		<kernel.version>1.2.1-java21-SNAPSHOT</kernel.version>
+		<maven.compiler.source>21</maven.compiler.source>
+ 		<maven.compiler.target>21</maven.compiler.target>
+		<jacoco.maven.plugin.version>0.8.11</jacoco.maven.plugin.version>
  		<maven.compiler.version>3.8.0</maven.compiler.version>
- 		<spring.boot.version>2.0.2.RELEASE</spring.boot.version>
- 		<java.version>11</java.version>
+ 		<java.version>21</java.version>
 	</properties>
+	<dependencyManagement>
+		<dependencies>
+			<dependency>
+				<groupId>io.mosip.kernel</groupId>
+				<artifactId>kernel-bom</artifactId>
+				<version>1.2.1-java21-SNAPSHOT</version>
+				<type>pom</type>
+				<scope>import</scope>
+			</dependency>
+
+		</dependencies>
+	</dependencyManagement>
 	<dependencies>
 		<dependency>
 		    <groupId>io.mosip.kernel</groupId>
 		    <artifactId>kernel-core</artifactId>
-		    <version>1.2.0.1-SNAPSHOT</version>
+		    <version>${kernel.version}</version>
+			<exclusions>
+				<exclusion>
+					<groupId>org.bouncycastle</groupId>
+					<artifactId>bcprov-jdk15on</artifactId>
+				</exclusion>
+				<exclusion>
+					<groupId>org.bouncycastle</groupId>
+					<artifactId>bcprov-jdk18on</artifactId>
+				</exclusion>
+			</exclusions>
+		</dependency>
+		<dependency>
+			<groupId>org.bouncycastle</groupId>
+			<artifactId>bcpkix-jdk15on</artifactId>
+			<version>1.68</version>
 		</dependency>
 		<dependency>
 			<groupId>io.mosip.kernel</groupId>
